% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRfunctions.R
\name{GRfit}
\alias{GRfit}
\title{Extract GR parameters from a dataset}
\usage{
GRfit(inputData, groupingColumns, GRtable = "none")
}
\arguments{
\item{inputData}{a data table consisting of the columns 'cell_count__time0', 'cell_count__ctrl', cell_count', 'concentration', and additional grouping columns.}

\item{groupingColumns}{columns of inputData that identify the experiment (i.e. which columns to not average over)}
}
\value{
a data table of GR metrics parameters (GR50, GRmax, etc.) as well as goodness of fit measures.
}
\description{
This function takes in a dataset with the columns 'cell_count__time0', 'cell_count__ctrl',
'cell_count', 'concentration', and additional grouping variables and calculates growth rate
inhibition (GR) metrics for each experiment in the dataset.
}
\details{
By default, only the table of GR metrics is returned. If GRtable equals 'both', then a list with two elements is returned: the table of GR metrics and a table of the initial data with an added column of GR values. If GRtable equals 'only', then only the table with GR values is returned.
}
\author{
Nicholas Clark
}
\seealso{
\code{drm}
}

